update workflow commands
